chore: release 5.7.0
Knowl installs as a Claude Code plugin, and four places that described the
product incorrectly now describe it correctly. No schema change, nothing
removed.

Minor rather than patch: a190190 adds a feature. Same reading as 5.6.0, which
cited 3.3.0 shipping a new MCP tool as a minor.

Carries the plugin manifest, the no-AI notice fix, the reference and README
rewrites, the re-recorded demo, and the five drift comments corrected in #144 --
all merged after the 5.6.0 release commit.

VERSION DRIFT IS NOW CHECKED RATHER THAN REMEMBERED. .claude-plugin/plugin.json
carries the version by hand and nothing kept it in step; it would have shipped
saying 5.6.0. That is the same shape as the demo Dockerfile which sat pinned to
3.2.2 while users installed 5.5.0, and as the five comments in #144 that still
described behaviour changed a week earlier. So check-lockfile-version.mjs became
check-version-sync.mjs and covers the manifest too, wired as
`npm run check:versions` in package.json and the CI step.

It caught the drift on this release -- the first one after it was written --
which is the argument for having it.
